Put AppLogger in its own file
