+++ x-pack/solutions/chat/plugins/workchat-app/server/utils/app_logger.ts
@@ -1,0 +1,13 @@
+import {Logger} from "@kbn/logging";
+
+export class AppLogger {
+  private static instance: Logger;
+
+  public static getInstance(): Logger {
+    return AppLogger.instance;
+  }
+
+  public static setInstance(logger: Logger) {
+    AppLogger.instance = logger;
+  }
+}
