+++ x-pack/solutions/chat/plugins/workchat-app/server/utils/index.ts
@@ -6,3 +6,4 @@
  */
 
 export { createBuilder } from './so_filters';
+export { AppLogger } from './app_logger';
